[BTAT-2248] Refactor core models into a core package
