+++ app/models/BusinessDetailsModel.scala
@@ -18,6 +18,7 @@ package models
 
 import java.time.LocalDate
 
+import models.core.{AccountingPeriodModel, AddressModel, CessationModel, ContactDetailsModel}
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{Json, Reads, _}
