+++ app/models/PropertyDetailsModel.scala
@@ -19,6 +19,7 @@ package models
 import java.time.LocalDate
 
 import models.CustomReads.readInt
+import models.core.{AccountingPeriodModel, CessationModel, ContactDetailsModel}
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{Json, Reads, _}
