+++ app/models/core/AccountingPeriodModel.scala
@@ -14,27 +14,21 @@
  * limitations under the License.
  */
 
-package models
+package models.core
 
 import java.time.LocalDate
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{Json, Reads, _}
 
-case class CessationModel(date: Option[LocalDate], reason: Option[String])
+case class AccountingPeriodModel(start: LocalDate, end: LocalDate)
 
-object CessationModel {
+object AccountingPeriodModel {
 
-  val desReads: Reads[CessationModel] = (
-    (__ \ "cessationDate").readNullable[LocalDate] and
-      (__ \ "cessationReason").readNullable[String]
-  )(CessationModel.apply _)
+  val desReads: Reads[AccountingPeriodModel] = (
+    (__ \ "accountingPeriodStartDate").read[LocalDate] and
+      (__ \ "accountingPeriodEndDate").read[LocalDate]
+  )(AccountingPeriodModel.apply _)
 
-  def cessation(date: Option[LocalDate], reason: Option[String]): Option[CessationModel] =
-    (date,reason) match{
-      case (None,None) => None
-      case _ => Some(CessationModel(date,reason))
-    }
-
-  implicit val format: Format[CessationModel] = Json.format[CessationModel]
+  implicit val format: Format[AccountingPeriodModel] = Json.format[AccountingPeriodModel]
 }
