+++ app/models/core/ContactDetailsModel.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package models
+package models.core
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{Json, Reads, _}
