+++ it/assets/BusinessDetailsIntegrationTestConstants.scala
@@ -19,7 +19,8 @@ package assets
 import java.time.LocalDate
 
 import assets.BaseIntegrationTestConstants.testMtdRef
-import models.{AccountingPeriodModel, AddressModel, BusinessDetailsModel, ContactDetailsModel}
+import models.core.{AccountingPeriodModel, AddressModel, ContactDetailsModel}
+import models.{AddressModel, BusinessDetailsModel, ContactDetailsModel}
 import play.api.libs.json.{JsValue, Json}
 
 object BusinessDetailsIntegrationTestConstants {
